+++ netports/__init__.py
@@ -1,12 +1,12 @@
 """netports"""
 
 from netports.interface4 import Interface4
-from netports.ip import IP_NAMES, IP_PORTS, iip, iip_all, iip_nip, sip, sip_all
+from netports.ip import IP_NAMES, IP_PORTS, iip, iip_nip, sip
 from netports.ports import inumbers, ranges, snumbers
 from netports.range import Range
 from netports.ranges import Ranges
-from netports.tcp import itcp_all, stcp_all, stcp, itcp
-from netports.vlan import ivlan_all, svlan_all, ivlan, ivlan_hpe, svlan, svlan_hpe
+from netports.tcp import stcp, itcp
+from netports.vlan import ivlan, svlan
 
 __all__ = [
     "IP_NAMES",
@@ -15,23 +15,15 @@ __all__ = [
     "Range",
     "Ranges",
     "iip",
-    "iip_all",
     "iip_nip",
     "inumbers",
     "itcp",
-    "itcp_all",
     "ivlan",
-    "ivlan_all",
-    "ivlan_hpe",
     "ranges",
     "sip",
-    "sip_all",
     "snumbers",
     "stcp",
-    "stcp_all",
     "svlan",
-    "svlan_all",
-    "svlan_hpe",
 ]
 
 __version__ = "0.2.0"