+++ tests/test__tcp.py
@@ -10,16 +10,21 @@ class Test(unittest.TestCase):
 
     def test_valid__itcp(self):
         """itcp()"""
-        for items, req in [
-            ("", []),
-            ([], []),
-            (1, [1]),
-            ([65535], [65535]),
-            ([1, 3, 4, 5], [1, 3, 4, 5]),
-            ("1,3-5", [1, 3, 4, 5]),
+        for kwargs, req in [
+            (dict(), []),
+            (dict(items=""), []),
+            (dict(items=[]), []),
+            (dict(items=1), [1]),
+            (dict(items="1"), [1]),
+            (dict(items=[1]), [1]),
+            (dict(items=[65535]), [65535]),
+            (dict(items=[5, 5, 1, 3, 4]), [1, 3, 4, 5]),
+            (dict(items="3-5,1,3-5,1"), [1, 3, 4, 5]),
+            (dict(all=True), list(range(1, 65536))),
+            (dict(items="1", all=True), list(range(1, 65536))),
         ]:
-            result = tcp.itcp(items)
-            self.assertEqual(result, req, msg=f"{items=}")
+            result = tcp.itcp(**kwargs)
+            self.assertEqual(result, req, msg=f"{kwargs=}")
 
     def test_invalid__itcp(self):
         """itcp()"""
@@ -34,24 +39,23 @@ class Test(unittest.TestCase):
             with self.assertRaises(error, msg=f"{items=}"):
                 tcp.itcp(items)
 
-    def test_valid__itcp_all(self):
-        """itcp_all()"""
-        result = tcp.itcp_all()
-        req = list(range(1, 65536))
-        self.assertEqual(result, req, msg="itcp_all")
-
     def test_valid__stcp(self):
         """stcp()"""
-        for items, req in [
-            ("", ""),
-            ([], ""),
-            (1, "1"),
-            ([65535], "65535"),
-            ([1, 3, 4, 5], "1,3-5"),
-            ("1,3-5", "1,3-5"),
+        for kwargs, req in [
+            (dict(), ""),
+            (dict(items=""), ""),
+            (dict(items=[]), ""),
+            (dict(items="1"), "1"),
+            (dict(items=1), "1"),
+            (dict(items=[1]), "1"),
+            (dict(items=[65535]), "65535"),
+            (dict(items=[5, 5, 1, 3, 4]), "1,3-5"),
+            (dict(items="3-5,1,3-5,1"), "1,3-5"),
+            (dict(all=True), "1-65535"),
+            (dict(items="1", all=True), "1-65535"),
         ]:
-            result = tcp.stcp(items)
-            self.assertEqual(result, req, msg=f"{items=}")
+            result = tcp.stcp(**kwargs)
+            self.assertEqual(result, req, msg=f"{kwargs=}")
 
     def test_invalid__stcp(self):
         """stcp()"""
@@ -65,12 +69,6 @@ class Test(unittest.TestCase):
         ]:
             with self.assertRaises(error, msg=f"{items=}"):
                 tcp.stcp(items)
-
-    def test_valid__stcp_all(self):
-        """stcp_all()"""
-        result = tcp.stcp_all()
-        req = "1-65535"
-        self.assertEqual(result, req, msg="stcp_all")
 
 
 if __name__ == "__main__":