+++ tests/test__vlan.py
@@ -2,7 +2,11 @@
 
 import unittest
 
+import dictdiffer  # type: ignore
+
 from netports import vlan
+
+ALL_VLANS = list(range(1, 4095))
 
 
 class Test(unittest.TestCase):
@@ -11,55 +15,54 @@ class Test(unittest.TestCase):
     def test_valid__ivlan(self):
         """ivlan()"""
         for kwargs, req in [
+            (dict(), []),
             (dict(items=""), []),
             (dict(items=[]), []),
             (dict(items=1), [1]),
+            (dict(items="1"), [1]),
+            (dict(items=[1]), [1]),
             (dict(items=[4094]), [4094]),
-            (dict(items=[1, 3, 4, 5]), [1, 3, 4, 5]),
-            (dict(items="1,3-5"), [1, 3, 4, 5]),
+            (dict(items=[5, 5, 1, 3, 4]), [1, 3, 4, 5]),
+            (dict(items="3-5,1,3-5,1"), [1, 3, 4, 5]),
+            (dict(all=True), ALL_VLANS),
+            (dict(items="1", all=True), ALL_VLANS),
+            (dict(items="1,3-5", platform="cisco"), [1, 3, 4, 5]),
+            (dict(items="1 3 to 5", platform="hpe"), [1, 3, 4, 5]),
             (dict(items="1,3-5", splitter=",", range_splitter="-"), [1, 3, 4, 5]),
             (dict(items="1 3 to 5", splitter=" ", range_splitter=" to "), [1, 3, 4, 5]),
         ]:
             result = vlan.ivlan(**kwargs)
             self.assertEqual(result, req, msg=f"{kwargs=}")
 
-    def test_valid__ivlan_hpe(self):
-        """ivlan_hpe()"""
-        for items, req in [
-            ([1, 3, 4, 5], [1, 3, 4, 5]),
-            ("1 3 to 5", [1, 3, 4, 5]),
-        ]:
-            result = vlan.ivlan_hpe(items)
-            self.assertEqual(result, req, msg=f"{items=}")
-
     def test_invalid__ivlan(self):
         """ivlan()"""
-        for items, error in [
-            ("-1", ValueError),
-            ("0", ValueError),
-            ("4095", ValueError),
-            ([0], ValueError),
-            ([-1], ValueError),
-            ([4095], ValueError),
+        for kwargs, error in [
+            (dict(items="-1"), ValueError),
+            (dict(items="0"), ValueError),
+            (dict(items="4095"), ValueError),
+            (dict(items=[0]), ValueError),
+            (dict(items=[-1]), ValueError),
+            (dict(items=[4095]), ValueError),
+            (dict(items="1,3-5", platform="hpe"), ValueError),
+            (dict(items="1 3 to 5", platform="cisco"), ValueError),
         ]:
-            with self.assertRaises(error, msg=f"{items=}"):
-                vlan.ivlan(items)
-
-    def test_valid__ivlan_all(self):
-        """ivlan_all()"""
-        result = vlan.ivlan_all()
-        req = list(range(1, 4095))
-        self.assertEqual(result, req, msg="ivlan_all")
+            with self.assertRaises(error, msg=f"{kwargs=}"):
+                vlan.ivlan(**kwargs)
 
     def test_valid__svlan(self):
         """svlan()"""
         for kwargs, req in [
+            (dict(), ""),
             (dict(items=""), ""),
             (dict(items=[]), ""),
             (dict(items=1), "1"),
+            (dict(items="1"), "1"),
+            (dict(items=[1]), "1"),
             (dict(items=[4094]), "4094"),
-            (dict(items=[1, 3, 4, 5]), "1,3-5"),
-            (dict(items="1,3-5"), "1,3-5"),
+            (dict(items=[5, 5, 1, 3, 4]), "1,3-5"),
+            (dict(items="3-5,1,3-5,1"), "1,3-5"),
+            (dict(items=[1, 3, 4, 5], platform="cisco"), "1,3-5"),
+            (dict(items=[1, 3, 4, 5], platform="hpe"), "1 3 to 5"),
             (dict(items=[1, 3, 4, 5], splitter=",", range_splitter="-"), "1,3-5"),
             (dict(items=[1, 3, 4, 5], splitter=" ", range_splitter=" to "), "1 3 to 5"),
         ]:
@@ -75,24 +78,28 @@ class Test(unittest.TestCase):
             ([0], ValueError),
             ([-1], ValueError),
             ([4095], ValueError),
+            (dict(items="1,3-5", platform="hpe"), ValueError),
+            (dict(items="1 3 to 5", platform="cisco"), ValueError),
         ]:
             with self.assertRaises(error, msg=f"{items=}"):
                 vlan.svlan(items)
 
-    def test_valid__svlan_hpe(self):
-        """svlan_hpe()"""
-        for items, req in [
-            ([1, 3, 4, 5], "1 3 to 5"),
-            ("1 3 to 5", "1 3 to 5"),
-        ]:
-            result = vlan.svlan_hpe(items)
-            self.assertEqual(result, req, msg=f"{items=}")
+    # =========================== helpers ============================
 
-    def test_valid__svlan_all(self):
-        """svlan_all()"""
-        result = vlan.svlan_all()
-        req = "1-4094"
-        self.assertEqual(result, req, msg="svlan_all")
+    def test_valid__update_splitters(self):
+        """_update_splitters()"""
+        for kwargs, req_d in [
+            (dict(), dict()),
+            (dict(splitter="a", range_splitter="a"), dict(splitter="a", range_splitter="a")),
+            (dict(platform="cisco"), dict(platform="cisco", splitter=",", range_splitter="-")),
+            (dict(platform="cisco", splitter="a", range_splitter="a"),
+             dict(platform="cisco", splitter=",", range_splitter="-")),
+            (dict(platform="hpe", splitter="a", range_splitter="a"),
+             dict(platform="hpe", splitter=" ", range_splitter=" to ")),
+        ]:
+            result = vlan._update_splitters(**kwargs)
+            diff = list(dictdiffer.diff(first=result, second=req_d))
+            self.assertEqual(diff, [], msg=f"{kwargs=}")
 
 
 if __name__ == "__main__":