+++ CHANGELOG.rst
@@ -5,7 +5,7 @@ CHANGELOG
 
 0.2.0 (2022-05-03)
 -------
-* [new] itcp_all, stcp_all, ivlan_all, svlan_all, iip, sip, iip_nip
+* [new] itcp(all=True), stcp(all=True), ivlan(all=True), svlan(all=True), iip, sip, iip_nip
 * [change] Ranges.ports > Ranges.numbers
 
 0.1.1 (2022-05-02)